test: move SetEnv()/UnSetEnv() to the testing_util library
We should never be modifying the environment from a service library,
so move `SetEnv()` and `UnSetEnv()` to `testing_util`, next to
`ScopedEnvironment`.

And while we're here, add warnings that these environment-modifying
functions should only be called when the process is single-threaded.
