+++ google/cloud/google_cloud_cpp_common.bzl
@@ -64,7 +64,6 @@ google_cloud_cpp_common_hdrs = [
     "internal/port_platform.h",
     "internal/random.h",
     "internal/retry_policy.h",
-    "internal/setenv.h",
     "internal/sha256_hash.h",
     "internal/status_payload_keys.h",
     "internal/strerror.h",
@@ -107,7 +106,6 @@ google_cloud_cpp_common_srcs = [
     "internal/populate_common_options.cc",
     "internal/random.cc",
     "internal/retry_policy.cc",
-    "internal/setenv.cc",
     "internal/sha256_hash.cc",
     "internal/status_payload_keys.cc",
     "internal/strerror.cc",
