+++ google/cloud/testing_util/google_cloud_cpp_testing.bzl
@@ -30,6 +30,7 @@ google_cloud_cpp_testing_hdrs = [
     "scoped_environment.h",
     "scoped_log.h",
     "scoped_thread.h",
+    "setenv.h",
     "status_matchers.h",
     "testing_types.h",
     "timer.h",
@@ -42,6 +43,7 @@ google_cloud_cpp_testing_srcs = [
     "mock_fake_clock.cc",
     "scoped_environment.cc",
     "scoped_log.cc",
+    "setenv.cc",
     "status_matchers.cc",
     "testing_types.cc",
     "timer.cc",
