+++ google/cloud/testing_util/setenv.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_SETENV_H
-#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_SETENV_H
+#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_TESTING_UTIL_SETENV_H
+#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_TESTING_UTIL_SETENV_H
 
 #include "google/cloud/version.h"
 #include "absl/types/optional.h"
@@ -22,10 +22,7 @@
 namespace google {
 namespace cloud {
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
-namespace internal {
-
-/// Unset (remove) an environment variable.
-void UnsetEnv(char const* variable);
+namespace testing_util {
 
 /**
  * Set the @p variable environment variable to @p value.
@@ -37,6 +34,10 @@ void UnsetEnv(char const* variable);
  *
  * @see
  * https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/putenv-s-wputenv-s?view=vs-2019
+ *
+ * @warning A modification to the environment must be serialized with all
+ *   other environment reads and writes, so this should only be used when
+ *   we are single-threaded.
  */
 void SetEnv(char const* variable, char const* value);
 
@@ -44,12 +45,25 @@ void SetEnv(char const* variable, char const* value);
  * Set the @p variable environment variable to @p value.
  *
  * If @value is an unset `absl::optional` then the variable is unset.
+ *
+ * @warning A modification to the environment must be serialized with all
+ *   other environment reads and writes, so this should only be used when
+ *   we are single-threaded.
  */
 void SetEnv(char const* variable, absl::optional<std::string> value);
 
-}  // namespace internal
+/**
+ * Unset (remove) an environment variable.
+ *
+ * @warning A modification to the environment must be serialized with all
+ *   other environment reads and writes, so this should only be used when
+ *   we are single-threaded.
+ */
+void UnsetEnv(char const* variable);
+
+}  // namespace testing_util
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
 }  // namespace cloud
 }  // namespace google
 
-#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_SETENV_H
+#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_TESTING_UTIL_SETENV_H
